+++ src/main/java/nordmods/iobvariantloader/util/dragon_variant_spawner/DragonVariantSpawnerReloadListener.java
@@ -9,6 +9,8 @@ import net.minecraft.server.packs.resources.ResourceManager;
 import net.minecraft.server.packs.resources.SimpleJsonResourceReloadListener;
 import net.minecraft.util.GsonHelper;
 import net.minecraft.util.profiling.ProfilerFiller;
+import nordmods.iobvariantloader.IoBVariantLoader;
+import nordmods.iobvariantloader.util.ResourceUtil;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.ArrayList;
@@ -25,11 +27,16 @@ public class DragonVariantSpawnerReloadListener extends SimpleJsonResourceReload
     protected void apply(Map<ResourceLocation, JsonElement> map, @NotNull ResourceManager pResourceManager, @NotNull ProfilerFiller pProfiler) {
         DragonVariantSpawnerUtil.dragonVariants.clear();
         for (Map.Entry<ResourceLocation, JsonElement> entry : map.entrySet()) {
-            String dragon = entry.getKey().getPath();
+            JsonObject entryObject = entry.getValue().getAsJsonObject();
+
+            String dragon = entryObject.has("dragon") ? entryObject.get("dragon").getAsString() : entry.getKey().getPath();
+            if (!ResourceUtil.AllowedValues.isValid(dragon, true)) {
+                IoBVariantLoader.LOGGER.warn("Variant spawns entry {} does not match any dragon id and will be skipped", entry.getKey());
+                continue;
+            }
             List<DragonVariantSpawner> variants = new ArrayList<>();
 
-            JsonArray array = entry.getValue().getAsJsonObject().get("variants").getAsJsonArray();
-
+            JsonArray array = entryObject.get("variants").getAsJsonArray();
             for (JsonElement elem : array) {
                 JsonObject input = elem.getAsJsonObject();
                 String name = input.get("name").getAsString();